AHCI CD-ROM Support.(SATAPI)

+++ user/lib/unistd.c
@@ -53,7 +53,7 @@ __syscall1(int,open,const char *,filename);
 __syscall1(int,close,int,fd);
 __syscall1(int,reboot,unsigned long,command);
 __syscall3(int,execve,const char *,path,const char **,argc,const char **,envp);
-__syscall3(int,read,int,fd,void *,buf,unsigned long,size);
-__syscall3(int,write,int,fd,const void *,buf,unsigned long,size);
+__syscall3(unsigned long,read,int,fd,void *,buf,unsigned long,size);
+__syscall3(unsigned long,write,int,fd,const void *,buf,unsigned long,size);
 __syscall3(int,waitpid,int,pid,int *,result,int,nowait);
